refactor: add re-export
