+++ src/components/core.js
@@ -1,0 +1,4 @@
+export * from './app/app.jsx';
+export * from './main/main.jsx';
+export * from './movies-list/movies-list.jsx';
+export * from './small-movie-card/small-movie-card.jsx';
