+++ src/components/small-movie-card/small-movie-card.test.js
@@ -1,6 +1,6 @@
 import React from "react";
 import renderer from "react-test-renderer";
-import SmallMovieCard from "./small-movie-card";
+import {SmallMovieCard} from "./small-movie-card";
 import {moviesData} from "../../mocks/films";
 
 it(`SmallMovieCard is rendered`, () => {
