+++ src/index.js
@@ -1,6 +1,6 @@
 import React from "react";
 import ReactDom from "react-dom";
-import App from "./components/app/app";
+import {App} from "./components/core";
 
 import {moviesData} from "./mocks/films";
 import {promoData} from "./mocks/promo";
